nf_tables: Update configuration and make it more modular
The current nf_tables configuration is outdated and enables modules that
no longer exist for current versions of the kernel. Also, it enables all
the nf_tables configuration options, there is no modularity.

This will update the configuration options for the current kernels and
make the kernel features more modular.

Signed-off-by: Mariano López <just.another.mariano@gmail.com>